requirements.txt
streamlit==0.75.0
tqdm==4.47.0
numpy==1.18.5
torch==1.6.0
torchvision==0.7.0
Pillow==8.2.0
